requirements.txt
fastapi==0.128.0
uvicorn==0.23.1
passlib[bcrypt]==1.7.4
bcrypt>=4.0.1
PyJWT==2.8.0
pytz==2025.2
email-validator==2.3.0
python-multipart==0.0.21









